ThreeSum防止int型溢出，加long即可

+++ src/txt/chap1/chap1_4/ThreeSum.java
@@ -28,7 +28,7 @@ public class ThreeSum {
         for (int i = 0; i < N; i++) {
             for (int j = i + 1; j < N; j++) {
                 for (int k = j + 1; k < N; k++) {
-                    if (a[i] + a[j] + a[k] == 0)
+                    if ((long)a[i] + a[j] + a[k] == 0)
                         cnt++;
                 }
             }
